drone_controller/requirements.txt
# Drone Controller Requirements
mavsdk>=2.0.0
asyncio
aiohttp>=3.9.0
pydantic>=2.0.0
python-dotenv>=1.0.0
structlog>=24.0.0
